# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba Tecnica Jeferson Mercedes",
        "description": "Documentación de la API",
        "contact": {
            "email": "jeferson.md0212@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products/{id}/prices": {
            "get": {
                "tags": [
                    "Precios de Productos"
                ],
                "summary": "Obtener los precios de un producto en diferentes monedas",
                "operationId": "003260c6ec8c5fef474b734a6aa7a706",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de precios obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "product_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "original_price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 100.5
                                            },
                                            "converted_price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 120.5
                                            },
                                            "currency": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Dólar"
                                                    },
                                                    "symbol": {
                                                        "type": "string",
                                                        "example": "$"
                                                    },
                                                    "exchange_rate": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1.2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Precios de Productos"
                ],
                "summary": "Crear un precio para un producto",
                "operationId": "bed1f398127625c2b3eccd89f348db35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Precio creado exitosamente"
                    },
                    "400": {
                        "description": "Petición incorrecta, faltan parámetros"
                    }
                }
            }
        }
    },
...
}